% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Plots.R
\docType{data}
\name{Basic_Plots}
\alias{Basic_Plots}
\title{Basic Plots}
\format{
A list with
\describe{
\item{\code{1}}{Number of Votes vs Average Rating}
\item{\code{2}}{Number of episodes in seasons and their runtime}
\item{\code{3}}{Average ratings and votes for each season}
}
}
\usage{
Basic_Plots
}
\description{
Average rating and votes summarised in different ways.
}
\examples{
length(Basic_Plots)

}
\keyword{datasets}
