% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SouthPark_Script_Data.R
\docType{data}
\name{SouthPark_Script_Data}
\alias{SouthPark_Script_Data}
\title{SouthPark Script Data}
\format{
A dataframe with
\describe{
\item{\code{Season}}{Season Number}
\item{\code{Episode}}{Episode Number}
\item{\code{Character}}{Character Name}
\item{\code{Line}}{The lines the character spoke}
}
}
\usage{
SouthPark_Script_Data
}
\description{
Data for the scripts scraped from the website are stored here.
The data consists of season, episode, character and line.
}
\examples{
unique(SouthPark_Script_Data$Season) # the seasons of the show
unique(SouthPark_Script_Data$Character) |> length() # the unique characters in the show

}
\keyword{datasets}
