% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Southpark_Summary.R
\docType{data}
\name{Southpark_Summary}
\alias{Southpark_Summary}
\title{SouthPark Summary}
\format{
A dataframe with
\describe{
\item{\code{Trivia}}{trivial information labels}
\item{\code{Values}}{data for the trivial information}
}
}
\usage{
Southpark_Summary
}
\description{
Overall summary plot from the script data.
}
\keyword{datasets}
