% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swear_Words_Plots.R
\docType{data}
\name{Swear_Words_Plots}
\alias{Swear_Words_Plots}
\title{Swear Words Plots}
\format{
A list with
\describe{
\item{\code{1}}{Swear word rate per 100 words in general}
\item{\code{2}}{Swear word rate per 100 words for main characters}
\item{\code{3}}{Swear word rate per 100 words for supporting characters}
}
}
\usage{
Swear_Words_Plots
}
\description{
Swear word plots for main and supporting characters per seasons.
Total number of words and unique words are summarised through plots.
}
\examples{
length(Swear_Words_Plots)

}
\keyword{datasets}
