% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition_Plots.R
\docType{data}
\name{Transition_Plots}
\alias{Transition_Plots}
\title{Transition Plots}
\format{
An object of class \code{ggraph} (inherits from \code{ggplot2::ggplot}, \code{ggplot}, \code{ggplot2::gg}, \code{S7_object}, \code{gg}) of length 1.
}
\usage{
Transition_Plots
}
\description{
A network showing speaking transitions between Cartman, Stan, Kyle
and Kenny across the selected seasons.
Nodes represent characters, displayed using their images.
An arrow from A to B indicates that B speaks immediately after A
within the same episode.
Arrow widths and labels show the percentage of transitions from A
to B, using all transitions from A to a different speaker as the denominator.
Only connections between the four selected characters are displayed.
Speaking order does not necessarily indicate whom a character addresses.
}
\examples{
length(Transition_Plots)

}
\keyword{datasets}
