% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_plots.R
\docType{data}
\name{Basic_plots}
\alias{Basic_plots}
\title{Basic plots}
\format{
A list with
\describe{
\item{\code{1}}{Number of Votes vs Average Rating}
\item{\code{2}}{Number of episodes in seasons and their runtime}
\item{\code{3}}{Average ratings and votes for each season}
}
}
\usage{
Basic_plots
}
\description{
Average rating and votes summarised in different ways.
}
\examples{
length(Basic_plots)

}
\keyword{datasets}
