% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_Grams_plots.R
\docType{data}
\name{N_Grams_plots}
\alias{N_Grams_plots}
\title{N Grams plots}
\format{
A list with
\describe{
\item{\code{1}}{Three word pharases over seasons}
\item{\code{2}}{Four word pharases over seasons}
\item{\code{3}}{Three word pharases over main characters}
\item{\code{4}}{Four word pharases over main characters}
\item{\code{5}}{Three word pharases over supporting characters}
\item{\code{6}}{Four word pharases over supporting characters}
}
}
\usage{
N_Grams_plots
}
\description{
Three and four word phrases common among seasons, main characters and supporting characters
are summarised through a plot here, from the script data.
}
\examples{
length(N_Grams_plots)

}
\keyword{datasets}
