% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratings_Votes_plots.R
\docType{data}
\name{Ratings_Votes_plots}
\alias{Ratings_Votes_plots}
\title{Ratings Votes plots}
\format{
A list with
\describe{
\item{\code{1}}{Rating for all seasons and episodes}
\item{\code{2}}{Votes for all seasons and episodes}
}
}
\usage{
Ratings_Votes_plots
}
\description{
Detailed plots for ratings and votes from the IMDB data.
}
\examples{
length(Ratings_Votes_plots)

}
\keyword{datasets}
