% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentiment_General_plots.R
\docType{data}
\name{Sentiment_General_plots}
\alias{Sentiment_General_plots}
\title{Sentiment General plots}
\format{
A list with
\describe{
\item{\code{1}}{positive words from bing over seasons}
\item{\code{2}}{negative words from bing over seasons}
\item{\code{3}}{positive words from nrc over seasons}
\item{\code{4}}{negative words from nrc over seasons}
\item{\code{5}}{positive words from loughran over seasons}
\item{\code{6}}{negative words from loughran over seasons}
}
}
\usage{
Sentiment_General_plots
}
\description{
Number of positive and negative words based on bing,nrc,loughran among the seasons
are summarised through a plot here, from the script data.
}
\examples{
length(Sentiment_General_plots)

}
\keyword{datasets}
