% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swear_Words_plots.R
\docType{data}
\name{Swear_Words_plots}
\alias{Swear_Words_plots}
\title{Swear words plots}
\format{
A list with
\describe{
\item{\code{1}}{Swear words in general}
\item{\code{2}}{Swear words for main characters}
\item{\code{3}}{Swear words for supporting characters}
}
}
\usage{
Swear_Words_plots
}
\description{
Swear word plots for main and supporting characters per seasons.
Total number of words and unique words are summarised through plots.
}
\examples{
length(Swear_Words_plots)

}
\keyword{datasets}
